Alterado componente detalhes para pagina e ajustado logout
