+++ src/pages/principal/Principal.js
@@ -10,8 +10,6 @@ import { useAuth } from '../../context/AuthContext';
 
 export default function({ navigation }) {
     const[dadosLista, setDados] = useState('');
-    const[detalhes, mostraDetalhes] = useState(false);
-    const[selecionado, setDetalhes] = useState('');
     const { Token, User } = useAuth();
 
     api.defaults.headers.common['Authorization'] = `Basic ${Token}`;
@@ -31,12 +29,11 @@ export default function({ navigation }) {
     function apresentaDetalhes(_idSelecionado) {               
         api.post(`/usuario/dadosSelecionado`, { _id : _idSelecionado })
             .then(response => {
-                setDetalhes(response.data);
+                navigation.navigate('Detalhes', { DadosSelecionado: response.data });
             })
             .catch(error => {
                 console.log(error);
             });
-        mostraDetalhes(!detalhes);
     }
 
     return(
@@ -51,8 +48,7 @@ export default function({ navigation }) {
                 <Image source={iconPesquisa} style={Styles.imagem}/>                 
             </View>
             <SafeAreaView style={Styles.formNavegacao}>
-                {detalhes == false &&
-                    <FlatList
+                <FlatList
                     showsVerticalScrollIndicator={false}
                     data={dadosLista}
                     keyExtractor={(item, dadosLista) => dadosLista.toString()}
@@ -67,14 +63,7 @@ export default function({ navigation }) {
                             </TouchableOpacity>          
                         )
                     }}
-                    />
-                }
-                {detalhes == true &&
-                    <DetalhesUsuario 
-                        sair={() => mostraDetalhes(!detalhes)} 
-                        UserSelecionado={selecionado} 
-                        UserLogado={User}/>
-                }
+                />
             </SafeAreaView>    
             
         </SafeAreaView>
